class mdsys of python wrapper
